changed the score datatype from integer to floating point

+++ app/judge/group/leaderboard/page.js
@@ -77,7 +77,7 @@ export default function GroupEventLeaderboardPage() {
                 group.judgeWiseTotal[judgeId] = 0;
               }
 
-              group.judgeWiseTotal[judgeId] += parseInt(
+              group.judgeWiseTotal[judgeId] += parseFloat(
                 group.score[_eventName][judgeId][criteria],
               );
             });
